+++ PgRoutiner/DataAccess/GetExtensions.cs
@@ -1,0 +1,20 @@
+﻿using System.Data;
+using Norm;
+using PgRoutiner.DataAccess.Models;
+
+namespace PgRoutiner.DataAccess;
+
+public static partial class DataAccessConnectionExtensions
+{
+    public static IEnumerable<PgItem> GetExtensions(this NpgsqlConnection connection)
+    {
+        return connection
+            .Read<string>("select extname from pg_extension")
+            .Select(s => new PgItem
+            {
+                Schema = null,
+                Name = s,
+                Type = PgType.Extension
+            });
+    }
+}