+++ PgRoutiner/DataAccess/Models/PgItem.cs
@@ -12,6 +12,10 @@ public static class PgItemExt
 {
     public static string GetFileName(this PgItem item)
     {
+        if (item.Schema == null)
+        {
+            return $"{item.Name}.sql";
+        }
         if (!string.Equals(item.Schema, "public"))
         {
             return $"{item.Schema}.{item.Name}.sql";