+++ PgRoutiner/DataAccess/Models/PgParameter.cs
@@ -7,4 +7,5 @@ public class PgParameter
     public string Type { get; set; }
     public string DataType { get; set; }
     public bool IsArray { get; set; }
+    public string Default { get; set; }
 }